# frontend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/transaction": {
      "get": {
        "description": "Возвращает список всех транзакций с пагинацией",
        "operationId": "TransactionController_getAllTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Максимальное количество записей",
            "schema": {
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список транзакций",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "total": 50,
                    "transactions": [
                      {
                        "id": 1,
                        "positionId": "5HauEoJa...",
                        "type": "OPEN_POSITION",
                        "date": "2025-01-15T10:30:00Z",
                        "txHash": "abc123...",
                        "baseToken": {
                          "symbol": "SOL",
                          "amount": 21.526289,
                          "valueUSD": 3996.54
                        },
                        "quoteToken": {
                          "symbol": "USDC",
                          "amount": 3842.7,
                          "valueUSD": 3842.7
                        },
                        "solPrice": 185.79,
                        "positionBalanceUSD": 7839.24,
                        "walletBalanceUSD": 248.88,
                        "profit": {
                          "usd": 125.5
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить историю транзакций",
        "tags": [
          "transaction"
        ]
      }
    },
    "/transaction/statistics": {
      "get": {
        "description": "Возвращает общую статистику по всем транзакциям",
        "operationId": "TransactionController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статистика",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "totalTransactions": 150,
                    "openPositions": 75,
                    "closePositions": 75,
                    "totalProfit": "1250.50",
                    "totalVolume": "125000.00"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить статистику транзакций",
        "tags": [
          "transaction"
        ]
      }
    },
    "/transaction/position/{positionId}": {
      "get": {
        "description": "Возвращает все транзакции для конкретной позиции",
        "operationId": "TransactionController_getPositionHistory",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "description": "ID позиции",
            "schema": {
              "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "История позиции",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "positionId": "5HauEoJa...",
                    "total": 2,
                    "transactions": [
                      {
                        "id": 1,
                        "type": "OPEN_POSITION",
                        "date": "2025-01-15T10:30:00Z"
                      },
                      {
                        "id": 2,
                        "type": "CLOSE_POSITION",
                        "date": "2025-01-16T14:20:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Позиция не найдена"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить историю позиции",
        "tags": [
          "transaction"
        ]
      }
    },
    "/transaction/grouped": {
      "get": {
        "description": "Возвращает список транзакций, где свапы сгруппированы по позициям. Свапы между Remove и Add Liquidity объединяются в SWAP_GROUP.",
        "operationId": "TransactionController_getGroupedTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Максимальное количество записей",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список транзакций с группировкой свапов",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "total": 25,
                    "transactions": [
                      {
                        "id": "swap-group-GFmD4Uc...",
                        "type": "SWAP_GROUP",
                        "positionId": "GFmD4UcWUTmnq...",
                        "swapCount": 3,
                        "date": "2025-12-12T19:51:25.351Z",
                        "swaps": [
                          {
                            "id": "0bce3b0d-18e0...",
                            "index": 1,
                            "txHash": "rPHW4i73aio9...",
                            "date": "2025-12-12T19:47:24.156Z",
                            "inputToken": "SOL",
                            "inputAmount": 12.903954322,
                            "inputValueUSD": 1705.27,
                            "outputToken": "USDC",
                            "outputAmount": 1684.196605,
                            "outputValueUSD": 1684.2,
                            "profitUSD": -21.08
                          }
                        ],
                        "totalProfitUSD": -23.92,
                        "solPrice": 131.3,
                        "walletBalanceUSD": 143.18
                      },
                      {
                        "id": "b6f730af-1c1d...",
                        "positionId": "GFmD4UcWUTmnq...",
                        "type": "OPEN_POSITION",
                        "date": "2025-12-12T19:47:40.468Z",
                        "txHash": "66APpa6qFfkf...",
                        "baseToken": {
                          "symbol": "SOL",
                          "amount": 13.715802339,
                          "valueUSD": 1792.06
                        },
                        "quoteToken": {
                          "symbol": "USDC",
                          "amount": 1955.810927,
                          "valueUSD": 1955.81
                        },
                        "solPrice": 131.16,
                        "positionBalanceUSD": 3747.87,
                        "walletBalanceUSD": 3797.57,
                        "profit": null
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить историю транзакций с группировкой свапов",
        "tags": [
          "transaction"
        ]
      }
    },
    "/transaction/monthly/{year}/{month}": {
      "get": {
        "description": "Возвращает чистую и среднюю прибыль за указанный месяц",
        "operationId": "TransactionController_getMonthlyProfit",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "example": 2025,
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "path",
            "description": "1-12",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика прибыли за месяц",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "stats": {
                      "year": 2025,
                      "month": 1,
                      "totalNetProfit": "$100.00",
                      "avgProfit": "$20.00",
                      "operations": 5,
                      "profitableOps": 4,
                      "lossOps": 1,
                      "successRate": "80.0%"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить прибыль за месяц",
        "tags": [
          "transaction"
        ]
      }
    },
    "/transaction/all-time": {
      "get": {
        "description": "Возвращает общую статистику прибыли и разбивку по месяцам",
        "operationId": "TransactionController_getAllTimeProfit",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статистика прибыли за все время",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "stats": {
                      "totalNetProfit": "$450.00",
                      "avgProfit": "$22.50",
                      "operations": 20,
                      "profitableOps": 17,
                      "lossOps": 3,
                      "successRate": "85.0%",
                      "monthlyBreakdown": [
                        {
                          "year": 2025,
                          "month": 3,
                          "profit": 150,
                          "operations": 8,
                          "avgProfit": 18.75
                        },
                        {
                          "year": 2025,
                          "month": 2,
                          "profit": 200,
                          "operations": 7,
                          "avgProfit": 28.57
                        },
                        {
                          "year": 2025,
                          "month": 1,
                          "profit": 100,
                          "operations": 5,
                          "avgProfit": 20
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить прибыль за все время",
        "tags": [
          "transaction"
        ]
      }
    },
...
  },
  "info": {
    "title": "Liquidity Bot API",
    "description": "API для управления ликвидностью на Raydium CLMM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}